Functions of JCAMPDXir

JCAMPDXir.JCAMPDXir — Module

This JCAMP-DX (infrared) file format was developed for the exchange of infrared spectra between different laboratories. General descripton of JCAMP-DX infrared format can be found in [https://iupac.org/wp-content/uploads/2021/08/JCAMP-DX_IR_1988.pdf] In addition to the spectra themselves, the file also stores metadata containing information about the units of measurement and the conditions under which the spectra were acquired. A detailed specification of the format is provided via the link. This package was written for a specific task and cannot be considered as a complete implementation of all JCAMP-DX format specification; however, it implements some of the basic functionality to read read_jdx_file and write write_jdx_file JCAMP-DX files according to the JCAMP-DX=4.24.

JCAMP file content example:

        ##TITLE=1 
        ##JCAMP-DX=4.24
        ##DATATYPE=INFRARED SPECTRUM
        ##DATE=2021/10/17
        ##TIME=11:30
        ##XUNITS=1/CM
        ##YUNITS=TRANSMITTANCE
        ##XFACTOR=1
        ##YFACTOR=0.00699183
        ##FIRSTX=0
        ##LASTX=15801.4671743
        ##FIRSTY=11746.3412893072
        ##MAXX=15801.4671743
        ##MINX=0
        ##MAXY=3.75371e+006
        ##MINY=4040.25
        ##NPOINTS=16384
        ##XYDATA=(X++(Y..Y))
        0 1680010 821286 2148133 1505245 1537124 1367661 1147725 1134981
        7.71603 1166853 1213186 1029828 1067595 1135904 1195128 1157134 1150556
        15.4321 1266743 1164401 1014224 1022338 999780 1138781 1208950 1161258
        .
        .
        ##END

JCAMPDXir.convert! — Method

convert!(x::AbstractArray,::sUnits{T},::sUnits{T}) where T

This functions can be used convert y and x units 

supported x-units names: NANOMETERS,NM,MKM,MICROMETERS,1/CM

supported y-units names: ABSORBANCE,A,T,A.U.,TRANSMITTANCE,R,KUBELKA-MUNK,ARBITRARY UNITS,REFLECTANCE

Example

julia> convert!([1,2,3],xUnits("MKM"),xUnits("1/cm"))) 

JCAMPDXir.parse_headers — Method

parse_headers(file::String)

Parses headers from JCAMP-DX file

JCAMPDXir.read! — Method

read!(jdx::JDXfile)

fills precreated JDXfile object

JCAMPDXir.read_jdx_file — Method

read_jdx_file(file_name::String)

Read JCAMP format file file_name - full file name,
Input: 
    file_name - full file name
returns named tuple with fields :
       x - coordinate (wavelength, wavenumber or other)
       y - data
       headers - dictionary in "String => value" format with 
       headers values

JCAMPDXir.write_jdx_file — Function

write_jdx_file(file_name,x::Vector{Float64},y::Vector{Float64},x_units::String="1/CM",
    y_units::String="TRANSMITTANCE"; kwargs...)

Saves infrared spectrum given as two vectors x and y to JCAMP-DX file file_name. Input vector should be of the same size. Currently the package suppots only (X++(Y..Y)) table data format. JCAMP-DX 4.24 demands 80 symbols per file line of Y-data and 88 total symbols per line. The y-vector is stored in eight columns, thus the total number of points should be a multiple of eight. All last mod(length(y),8) points of y dtaa will not be written to the file. It is preferable that all x-data is sorted in ascending order and spaced uniformly. If it is not the case, or if the units conversion is envolved, function will automatically interpolate and sort the data on uniformly spaced grid.

x_units - units of x data, must be one of NANOMETERS,NM,MKM,MICROMETERS,1/CM

y_units - units of y data, must be one of ABSORBANCE,A,T,A.U.,TRANSMITTANCE,R,KUBELKA-MUNK,ARBITRARY UNITS,REFLECTANCE

Further any keword arguments can be provided, all of them will be written to the head of the file. All keyword arguments appear in the file in uppercase.

Most impostant are

TITLE - the title of the file (it is always on top of the file)
XUNITS - x data units saved to file,  must be one of  NANOMETERS,NM,MKM,MICROMETERS,1/CM
YUNITS - y data units saved to file, must be one of ABSORBANCE,A,T,A.U.,TRANSMITTANCE,R,KUBELKA-MUNK,ARBITRARY UNITS,REFLECTANCE

If x_units are not equal to the key-word argument XUNITS than the function converts x-values before saving to file see xconvert!

If y_units are not equal to the key-word argument XUNITS than the function converts y-values before saving to file see yconvert!

Example

julia> using JCAMPDXir
julia> filename = joinpath(@__DIR__,"test.jdx")
julia> write_jdx_file(filename,[1,2,3,4,5,6,7,8],rand(8),"MKM","T",title = "new file",XUNIT="1/CM",YUNITS="KUBELKA-MUNK") 

JCAMPDXir.xconvert! — Method

xconvert!(x::AbstractArray,input_units::String,output_units::String)

Converts the values of `x` from `input_units` to `output_units`.
supported x-units names: NANOMETERS,NM,MKM,MICROMETERS,1/CM

Example

julia> xconvert!([1,2,3],"MKM",xUnits"1/cm")) 

JCAMPDXir.yconvert! — Method

yconvert!(y::AbstractArray,input_units::String,output_units::String)

Converts the values of `y` from `input_units` to `output_units`.
supported x-units names: NANOMETERS,NM,MKM,MICROMETERS,1/CM
All units can be written both in lower- and in uppercase, `T`,`R` and `A` stay for 
a shorthand for `TRANSMITTANCE`,`REFLECTANCE` and `ABSORBANCE`

Example

julia> yconvert!([1,2,3],"R","KUBELKA-MUNK"))